+++ .changeset/shy-tokens-arrive.md
@@ -1,6 +1,7 @@
 ---
+"@gradio/client": minor
 "gradio": minor
 "gradio_client": minor
 ---
 
-feat:oauth: let API callers supply a token for endpoints that take a `gr.OAuthToken`, via `Client(..., oauth_token=...)` — sent only to endpoints that declare they need one
+feat:oauth: let API callers supply a token for endpoints that take a `gr.OAuthToken`, via `oauth_token` on the Python and JS clients — sent only to endpoints that declare they need one
